Restructure filesystem (utils package)
