+++ server/app.py
@@ -1,15 +1,19 @@
 import json
 import logging
-import sys
 import traceback
-from io import StringIO
 
-from flask import Flask, jsonify, Response, send_file
+from flask import Flask, jsonify, send_file
 from flask_cors import CORS
 
 from app import interop, uav, ugv
-from errors import InvalidRequestError, InvalidStateError, GeneralError, ServiceUnavailableError
+from utils.errors import (
+    InvalidRequestError,
+    InvalidStateError,
+    GeneralError,
+    ServiceUnavailableError,
+)
 from groundstation import GroundStation
+from utils.logging_setup import LOG_STREAM, TELEM_STREAM
 
 log = logging.getLogger("werkzeug")
 log.setLevel(logging.ERROR)
@@ -25,72 +29,11 @@ app.register_blueprint(interop, url_prefix="/interop")
 app.register_blueprint(uav, url_prefix="/uav")
 app.register_blueprint(ugv, url_prefix="/ugv")
 
-
-def log_level(self, message, *args, **kwargs):
-    if self.isEnabledFor(logging.INFO + 5):
-        self._log(logging.INFO + 5, message, args, **kwargs)
-
-
-def log_root(message, *args, **kwargs):
-    logging.log(logging.INFO + 5, message, *args, **kwargs)
-
-
-logging.addLevelName(logging.INFO + 5, "IMPORTANT")
-setattr(logging, "IMPORTANT", logging.INFO + 5)
-setattr(logging.getLoggerClass(), "important", log_level)
-setattr(logging, "important", log_root)
-
 logger = logging.getLogger("groundstation")
-logger.setLevel(logging.DEBUG)
 
-autopilot = logging.getLogger("autopilot")
-autopilot.setLevel(logging.DEBUG)
-
-telemetry = logging.getLogger("telemetry")
-telemetry.setLevel(logging.INFO)
-
-formatter = logging.Formatter("[%(levelname)-9s] (%(name)-13s) %(asctime)s  %(message)-500s")
-telem_formatter = logging.Formatter("%(asctime)s;%(message)s")
-
-# console_handler = logging.StreamHandler(sys.stdout)
-# console_handler.setLevel(logging.IMPORTANT)
-# console_handler.setFormatter(formatter)
-# logger.addHandler(console_handler)
-
-file_handler = logging.FileHandler("logs/info.log", mode="w")
-file_handler.setLevel(logging.INFO)
-file_handler.setFormatter(formatter)
-logger.addHandler(file_handler)
-autopilot.addHandler(file_handler)
-
-debug_file_handler = logging.FileHandler("logs/debug.log", mode="w")
-debug_file_handler.setLevel(logging.DEBUG)
-debug_file_handler.setFormatter(formatter)
-logger.addHandler(debug_file_handler)
-autopilot.addHandler(debug_file_handler)
-
-telem_file_handler = logging.FileHandler("logs/telem.log", mode="w")
-telem_file_handler.setLevel(logging.INFO)
-telem_file_handler.setFormatter(telem_formatter)
-telemetry.addHandler(telem_file_handler)
-
-LOG_STREAM = StringIO()
-string_handler = logging.StreamHandler(LOG_STREAM)
-string_handler.setLevel(logging.INFO)
-string_handler.setFormatter(formatter)
-logger.addHandler(string_handler)
-autopilot.addHandler(string_handler)
-
-TELEM_STREAM = StringIO()
-telem_string_handler = logging.StreamHandler(TELEM_STREAM)
-telem_string_handler.setLevel(logging.INFO)
-telem_string_handler.setFormatter(telem_formatter)
-telemetry.addHandler(telem_string_handler)
-
-logger.info("STARTED LOGGING")
-
-gs = GroundStation()
+gs = GroundStation(config=config)
 app.gs = gs
+app.config = config
 
 
 @app.errorhandler(Exception)
