+++ server/app/interop.py
@@ -3,7 +3,7 @@ import json
 
 from flask import Blueprint, current_app as app, request
 
-from errors import InvalidRequestError
+from utils.errors import InvalidRequestError
 
 interop = Blueprint("interop", __name__)
 
@@ -66,9 +66,7 @@ def odlc_filter(status):
 
 @interop_odlc.route("/image/<int:id_>")
 def odlc_get_image(id_):
-    with open("config.json", "r", encoding="utf-8") as f:
-        config = json.load(f)
-    file_extension = "jpg" if config["uav"]["images"]["quality"] > 0 else "png"
+    file_extension = "jpg" if app.config["uav"]["images"]["quality"] > 0 else "png"
     with open(f"assets/odlc_images/{id_}.{file_extension}", "rb") as image_file:
         encoded_string = base64.b64encode(image_file.read())
     return {"image": encoded_string.decode("utf-8")}
