+++ server/handlers/interop/dummy.py
@@ -6,12 +6,15 @@ import os
 import typing
 from datetime import datetime, timedelta, date
 
-from auvsi_suas.client import client
 from auvsi_suas.proto import interop_api_pb2 as interop
-from google.protobuf import json_format
 from requests.exceptions import ConnectionError as RequestsCE
 
-from errors import InvalidRequestError, InvalidStateError, GeneralError, ServiceUnavailableError
+from utils.errors import (
+    InvalidRequestError,
+    InvalidStateError,
+    GeneralError,
+    ServiceUnavailableError,
+)
 from handlers.utils import decorate_all_functions, log
 
 if typing.TYPE_CHECKING:
