+++ server/sample.config.json
@@ -20,6 +20,7 @@
         }
     },
     "interop": {
+        "type": "dummy",
         "url": "http://localhost:8000",
         "username": "testuser",
         "password": "testpass",
