+++ server/utils/clean_directory.py
